+++ tools/decode-status.py
@@ -259,7 +259,7 @@ a_features = [[
     "USE_BM8563","USE_ENERGY_DUMMY","USE_AM2320","USE_T67XX",
     "USE_MCP2515","USE_TASMESH","USE_WIFI_RANGE_EXTENDER","USE_INFLUXDB",
     "USE_HRG15","USE_VINDRIKTNING","USE_SCD40","USE_HM330X",
-    "","","","",
+    "USE_HDC2010","","","",
     "","","","",
     "","","","",
     "","","",""
@@ -290,7 +290,7 @@ else:
         obj = json.load(fp)
 
 def StartDecode():
-    print ("\n*** decode-status.py v20211008 by Theo Arends and Jacek Ziolkowski ***")
+    print ("\n*** decode-status.py v20211111 by Theo Arends and Jacek Ziolkowski ***")
 
 #    print("Decoding\n{}".format(obj))